# rest/api/v1/openapi.yml
...
openapi: 3.1.0
paths:
...
  /api/v1/network/exchangerate:
    get:
      summary: Get the network exchange rate to estimate costs
      description: Returns the network's exchange rate, current and next.
      operationId: getNetworkExchangeRate
      parameters:
        - $ref: "#/components/parameters/timestampQueryParam"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NetworkExchangeRateSetResponse"
        400:
          $ref: "#/components/responses/InvalidParameterError"
        404:
          $ref: "#/components/responses/NotFoundError"
        500:
          $ref: "#/components/responses/ServiceUnavailableError"
      tags:
        - network
  /api/v1/network/fees:
    get:
      summary: Get the network fees
      description: Returns the estimated gas in tinybars per each transaction type. Default order is ASC. Currently only `ContractCall`, `ContractCreate` and `EthereumTransaction` transaction types are supported.
      operationId: getNetworkFees
      parameters:
        - $ref: "#/components/parameters/orderQueryParam"
        - $ref: "#/components/parameters/timestampQueryParam"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NetworkFeesResponse"
        400:
          $ref: "#/components/responses/InvalidParameterError"
        404:
          $ref: "#/components/responses/NotFoundError"
        500:
          $ref: "#/components/responses/ServiceUnavailableError"
      tags:
        - network
    post:
      summary: Estimate network fees
      description: Given a HAPI transaction, estimate the network fees in tinycents.
      operationId: estimateFees
      parameters:
        - $ref: "#/components/parameters/estimateModeQueryParam"
        - $ref: "#/components/parameters/highVolumeThrottleQueryParam"
      requestBody:
        content:
          application/protobuf:
            schema:
              format: binary
              type: string
          application/x-protobuf:
            schema:
              format: binary
              type: string
        description: A protobuf encoded HAPI Transaction
        required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeeEstimateResponse"
        400:
          $ref: "#/components/responses/InvalidParameterError"
        500:
          $ref: "#/components/responses/ServiceUnavailableError"
      tags:
        - network
  /api/v1/network/nodes:
    get:
      summary: Get the network address book nodes
      description: Returns the network's list of nodes used in consensus
      operationId: getNetworkNodes
      parameters:
        - $ref: "#/components/parameters/fileIdQueryParam"
        - $ref: "#/components/parameters/limitQueryParam"
        - $ref: "#/components/parameters/nodeIdQueryParam"
        - $ref: "#/components/parameters/orderQueryParam"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NetworkNodesResponse"
        400:
          $ref: "#/components/responses/InvalidParameterError"
      tags:
        - network
  /api/v1/network/registered-nodes:
    get:
      summary: Get registered nodes
      description: Returns the list of registered nodes
      operationId: getRegisteredNodes
      parameters:
        - $ref: "#/components/parameters/limitQueryParam"
        - $ref: "#/components/parameters/orderQueryParam"
        - $ref: "#/components/parameters/registeredNodeIdQueryParam"
        - $ref: "#/components/parameters/registeredNodeTypeQueryParam"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisteredNodesResponse"
        400:
          $ref: "#/components/responses/InvalidParameterError"
      tags:
        - network
  /api/v1/network/stake:
    get:
      summary: Get network stake information
      description: Returns the network's current stake information.
      operationId: getNetworkStake
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NetworkStakeResponse"
        400:
          $ref: "#/components/responses/InvalidParameterError"
        404:
          $ref: "#/components/responses/NetworkStakeNotFound"
        500:
          $ref: "#/components/responses/ServiceUnavailableError"
      tags:
        - network
  /api/v1/network/supply:
    get:
      summary: Get the network supply
      description: Returns the network's released supply of hbars
      operationId: getNetworkSupply
      parameters:
        - $ref: "#/components/parameters/timestampQueryParam"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NetworkSupplyResponse"
        400:
          $ref: "#/components/responses/InvalidParameterError"
        404:
          $ref: "#/components/responses/NotFoundError"
      tags:
        - network
...
info:
  title: Mirror Node REST API
  version: 0.164.0-SNAPSHOT
  license:
    name: Apache-2.0
    url: "https://www.apache.org/licenses/LICENSE-2.0.html"
  description: "The REST API offers the ability to query transactions and entity information from a mirror node.\n\nBase url: [/api/v1](/api/v1)\n\nOpenAPI Spec: [/api/v1/docs/openapi.yml](/api/v1/docs/openapi.yml)"
  contact:
    name: Mirror Node Team
    email: mirrornode@hedera.com
    url: "https://github.com/hiero-ledger/hiero-mirror-node"
...
servers:
  - description: The current REST API server
    url: ""
  - description: The production REST API servers
    url: "{scheme}://{network}.mirrornode.hedera.com"
    variables:
      scheme:
        default: https
        description: The URI scheme
        enum: [http, https]
      network:
        default: testnet
        description: The Hedera network in use
        enum: [mainnet-public, mainnet, previewnet, testnet]
components:
  schemas:
...
    NetworkExchangeRateSetResponse:
      type: object
      properties:
        current_rate:
          $ref: "#/components/schemas/ExchangeRate"
        next_rate:
          $ref: "#/components/schemas/ExchangeRate"
        timestamp:
          $ref: "#/components/schemas/Timestamp"
    NetworkFeesResponse:
      type: object
      properties:
        fees:
          $ref: "#/components/schemas/NetworkFees"
        timestamp:
          $ref: "#/components/schemas/Timestamp"
    NetworkNodesResponse:
      type: object
      required:
        - nodes
        - links
      properties:
        nodes:
          $ref: "#/components/schemas/NetworkNodes"
        links:
          $ref: "#/components/schemas/Links"
    NetworkSupplyResponse:
      type: object
      properties:
        released_supply:
          description: The network's released supply of hbars in tinybars
          example: "3999999999999999949"
          type: string
        timestamp:
          allOf:
            - $ref: "#/components/schemas/Timestamp"
            - description: The consensus timestamp at which the released supply was valid
        total_supply:
          description: The network's total supply of hbars in tinybars
          example: "5000000000000000000"
          type: string
...
    RegisteredNodesResponse:
      type: object
      required:
        - links
        - registered_nodes
      properties:
        registered_nodes:
          $ref: "#/components/schemas/RegisteredNodes"
        links:
          $ref: "#/components/schemas/Links"
...
    EntityIdQuery:
      type: string
      pattern: ^((gte?|lte?|eq|ne)\:)?(\d{1,10}\.\d{1,10}\.)?\d{1,10}$
    Error:
      type: object
      properties:
        _status:
          type: object
          properties:
            messages:
              type: array
              items:
                type: object
                properties:
                  data:
                    description: Error message in hexadecimal
                    example: "0x3000"
                    format: binary
                    pattern: "^0x[0-9a-fA-F]+$"
                    type: [string, "null"]
                  detail:
                    description: Detailed error message
                    example: Generic detailed error message
                    type: [string, "null"]
                  message:
                    description: Error message
                    example: Generic error message
                    type: string
...
    ExchangeRate:
      type: object
      properties:
        cent_equivalent:
          format: int32
          example: 596987
          type: integer
        expiration_time:
          format: int64
          example: 1649689200
          type: integer
        hbar_equivalent:
          example: 30000
          format: int32
          type: integer
...
    Links:
      type: object
      properties:
        next:
          example: null
          type: [string, "null"]
...
    NetworkNodes:
      type: array
      items:
        $ref: "#/components/schemas/NetworkNode"
...
    NetworkFees:
      type: array
      items:
        $ref: "#/components/schemas/NetworkFee"
    NetworkStakeResponse:
      type: object
      required:
        - max_stake_rewarded
        - max_staking_reward_rate_per_hbar
        - max_total_reward
        - node_reward_fee_fraction
        - reserved_staking_rewards
        - reward_balance_threshold
        - stake_total
        - staking_period
        - staking_period_duration
        - staking_periods_stored
        - staking_reward_fee_fraction
        - staking_reward_rate
        - staking_start_threshold
        - unreserved_staking_reward_balance
      properties:
        max_stake_rewarded:
          description: |
            The maximum amount of tinybar that can be staked for reward while still achieving
            the maximum per-hbar reward rate
          format: int64
          type: integer
        max_staking_reward_rate_per_hbar:
          description: The maximum reward rate, in tinybars per whole hbar, that any account can receive in a day
          type: integer
          format: int64
        max_total_reward:
          description: |
            The total tinybars to be paid as staking rewards in the ending period,
            after applying the settings for the 0.0.800 balance threshold and the maximum stake rewarded
          format: int64
          type: integer
        node_reward_fee_fraction:
          description: The fraction between zero and one of the network and service fees paid to the node reward account 0.0.801
          type: number
          format: float
        reserved_staking_rewards:
          description: |
            The amount of the staking reward funds of account 0.0.800 reserved to pay pending
            rewards that have been earned but not collected
          format: int64
          type: integer
        reward_balance_threshold:
          description: |
            The unreserved tinybar balance of account 0.0.800 required to achieve the
            maximum per-hbar reward rate
          format: int64
          type: integer
        stake_total:
          description: The total amount staked to the network in tinybars the start of the current staking period
          type: integer
          format: int64
        staking_period:
          allOf:
            - $ref: "#/components/schemas/TimestampRange"
            - description: The timestamp range of the staking period
        staking_period_duration:
          description: The number of minutes in a staking period
          type: integer
          format: int64
        staking_periods_stored:
          description: The number of staking periods for which the reward is stored for each node
          type: integer
          format: int64
        staking_reward_fee_fraction:
          description: The fraction between zero and one of the network and service fees paid to the staking reward account 0.0.800
          type: number
          format: float
        staking_reward_rate:
          description: The total number of tinybars to be distributed as staking rewards each period
          type: integer
          format: int64
        staking_start_threshold:
          description: The minimum balance of staking reward account 0.0.800 required to active rewards
          type: integer
          format: int64
        unreserved_staking_reward_balance:
          description: |
            The unreserved balance of account 0.0.800 at the close of the just-ending period;
            this value is used to compute the HIP-782 balance ratio
          format: int64
          type: integer
      example:
        max_stake_rewarded: 10
        max_staking_reward_rate_per_hbar: 17808
        max_total_reward: 20
        node_reward_fee_fraction: 1.0
        reserved_staking_rewards: 30
        reward_balance_threshold: 40
        stake_total: 35000000000000000
        staking_period:
          from: "1655164800.000000000"
          to: "1655251200.000000000"
        staking_period_duration: 1440
        staking_periods_stored: 365
        staking_reward_fee_fraction: 1.0
        staking_reward_rate: 100000000000
        staking_start_threshold: 25000000000000000
        unreserved_staking_reward_balance: 50
...
    FeeEstimate:
      description: The fee estimate for the network component. Includes the base fee and any extras associated with it.
      properties:
        base:
          description: The base fee price, in tinycents.
          example: 1000
          format: int64
          minimum: 0
          type: integer
        extras:
          description: The extra fees that apply for this fee component.
          items:
            $ref: "#/components/schemas/FeeExtra"
          type: array
      required:
        - base
        - extras
      type: object
    FeeEstimateMode:
      description: Estimate solely based on the transaction's inherent properties or use network state.
      default: INTRINSIC
      enum: [INTRINSIC, STATE]
      type: string
    FeeEstimateNetwork:
      description: The network fee component which covers the cost of gossip, consensus, signature verifications, fee payment, and storage.
      properties:
        multiplier:
          description: Multiplied by the node fee to determine the total network fee.
          example: 2
          format: int32
          minimum: 0
          type: integer
        subtotal:
          description: The subtotal in tinycents for the network fee component which is calculated by multiplying the node subtotal by the network multiplier.
          example: 1200
          format: int64
          minimum: 0
          type: integer
      required:
        - multiplier
        - subtotal
      type: object
    FeeEstimateResponse:
      description: The response containing the estimated transaction fees.
      properties:
        high_volume_multiplier:
          description: The high-volume pricing multiplier per HIP-1313. A value of 1 indicates no high-volume pricing. A value greater than 1 applies when the transaction's highVolume flag is true and throttle utilization is non-zero.
          example: 100
          format: int64
          minimum: 1
          type: integer
        network:
          $ref: "#/components/schemas/FeeEstimateNetwork"
        node:
          allOf:
            - $ref: "#/components/schemas/FeeEstimate"
            - description: The node fee component which is to be paid to the node that submitted the transaction to the network. This fee exists to compensate the node for the work it performed to pre-check the transaction before submitting it, and incentivizes the node to accept new transactions from users.
        service:
          allOf:
            - $ref: "#/components/schemas/FeeEstimate"
            - description: The service fee component which covers execution costs, state saved in the Merkle tree, and additional costs to the blockchain storage.
        total:
          description: The sum of the network, node, and service subtotals in tinycents.
          example: 1000
          format: int64
          minimum: 0
          type: integer
      required:
        - high_volume_multiplier
        - network
        - node
        - service
        - total
      type: object
...
    RegisteredNodes:
      type: array
      items:
        $ref: "#/components/schemas/RegisteredNode"
    RegisteredNodeType:
      description: Registered node type.
      enum: [BLOCK_NODE, GENERAL_SERVICE, MIRROR_NODE, RPC_RELAY]
      example: BLOCK_NODE
      type: string
...
    Timestamp:
      description: A Unix timestamp in seconds.nanoseconds format
      type: string
      example: "1586567700.453054000"
      pattern: '^\d{1,10}(\.\d{1,9})?$'
...
    TimestampRange:
      type: object
      description: A timestamp range an entity is valid for
      properties:
        from:
          allOf:
            - $ref: "#/components/schemas/Timestamp"
            - description: The inclusive from timestamp in seconds
        to:
          allOf:
            - $ref: "#/components/schemas/TimestampNullable"
            - description: The exclusive to timestamp in seconds
...
  responses:
    NotFoundError:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            _status:
              messages:
                - message: Not found
    NetworkStakeNotFound:
      description: No network stake data found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            _status:
              messages:
                - detail: No network stake data found
                  message: Not Found
...
    InvalidParameterError:
      description: Invalid parameter
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            _status:
              messages:
                - message: "Invalid parameter: account.id"
                - message: Invalid Transaction id. Please use \shard.realm.num-sss-nnn\ format where sss are seconds and nnn are nanoseconds
    ServiceUnavailableError:
      description: Service Unavailable
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            _status:
              messages:
                - message: Require at least 1/3 signature files to prove consensus, got 1 out of 4 for file 2019-10-11T13_33_25.526889Z.rcd_sig
  parameters:
...
    fileIdQueryParam:
      name: file.id
      description: The ID of the file entity
      in: query
      examples:
        noValue:
          summary: --
          value: ""
        entityNumNoOperator:
          summary: Example of entityNum equals with no operator
          value: 100
        idNoOperator:
          summary: Example of id equals with no operator
          value: 0.0.100
        entityNumEqOperator:
          summary: Example of entityNum equals operator
          value: eq:200
        idEqOperator:
          summary: Example of id equals operator
          value: eq:0.0.200
      schema:
        $ref: "#/components/schemas/EntityIdQuery"
...
    limitQueryParam:
      name: limit
      in: query
      description: The maximum number of items to return
      example: 2
      schema:
        format: int32
        type: integer
        default: 25
        minimum: 1
        maximum: 100
...
    nodeIdQueryParam:
      name: node.id
      description: The ID of the node
      in: query
      examples:
        noValue:
          summary: --
          value: ""
        serialNumNoOperator:
          summary: Example of nodeId equals with no operator
          value: 1
        serialNumEqOperator:
          summary: Example of nodeId equals operator
          value: eq:2
        serialNumGtOperator:
          summary: Example of nodeId greater than operator
          value: gt:3
        serialNumGteOperator:
          summary: Example of nodeId greater than or equals operator
          value: gte:4
        serialNumLtOperator:
          summary: Example of nodeId less than operator
          value: lt:5
        serialNumLteOperator:
          summary: Example of nodeId less than or equals operator
          value: lte:6
      schema:
        type: string
        pattern: ^((eq|gt|gte|lt|lte):)?\d{1,19}$
...
    orderQueryParam:
      name: order
      in: query
      description: The order in which items are listed
      example: desc
      schema:
        enum: [asc, desc]
        default: asc
...
    registeredNodeIdQueryParam:
      name: registerednode.id
      description: The ID of the registered node
      in: query
      examples:
        noValue:
          summary: --
          value: ""
        serialNumNoOperator:
          summary: Example of registeredNodeId equals with no operator
          value: 1
        serialNumEqOperator:
          summary: Example of registeredNodeId equals operator
          value: eq:2
        serialNumGtOperator:
          summary: Example of registeredNodeId greater than operator
          value: gt:3
        serialNumGteOperator:
          summary: Example of registeredNodeId greater than or equals operator
          value: gte:4
        serialNumLtOperator:
          summary: Example of registeredNodeId less than operator
          value: lt:5
        serialNumLteOperator:
          summary: Example of registeredNodeId less than or equals operator
          value: lte:6
      schema:
        type: string
        pattern: ^((eq|gt|gte|lt|lte):)?\d{1,19}$
...
    timestampQueryParam:
      description: The consensus timestamp as a Unix timestamp in seconds.nanoseconds format with an optional comparison operator. See [unixtimestamp.com](https://www.unixtimestamp.com/) for a simple way to convert a date to the 'seconds' part of the Unix time.
      name: timestamp
      in: query
      explode: true
      examples:
        noValue:
          summary: --
          value: ""
        secondsNoOperator:
          summary: Example of seconds equals with no operator
          value: 1234567890
        timestampNoOperator:
          summary: Example of timestamp equals with no operator
          value: 1234567890.000000100
        secondsEqOperator:
          summary: Example of seconds equals with operator
          value: eq:1234567890
        timestampEqOperator:
          summary: Example of timestamp equals with operator
          value: eq:1234567890.000000200
        secondsNeOperator:
          summary: Example of seconds not equals operator
          value: ne:1234567890
        timestampNeOperator:
          summary: Example of timestamp not equals operator
          value: ne:1234567890.000000300
        secondsGtOperator:
          summary: Example of seconds greater than operator
          value: gt:1234567890
        timestampGtOperator:
          summary: Example of timestamp greater than operator
          value: gt:1234567890.000000400
        secondsGteOperator:
          summary: Example of seconds greater than or equals operator
          value: gte:1234567890
        timestampGteOperator:
          summary: Example of timestamp greater than or equals operator
          value: gte:1234567890.000000500
        secondsLtOperator:
          summary: Example of seconds less than operator
          value: lt:1234567890
        timestampLtOperator:
          summary: Example of timestamp less than operator
          value: lt:1234567890.000000600
        secondsLteOperator:
          summary: Example of seconds less than or equals operator
          value: lte:1234567890
        timestampLteOperator:
          summary: Example of timestamp less than or equals operator
          value: lte:1234567890.000000700
      schema:
        type: array
        items:
          type: string
          pattern: ^((eq|gt|gte|lt|lte|ne):)?\d{1,10}(\.\d{1,9})?$
...
    estimateModeQueryParam:
      description: Estimate solely based on the transaction's inherent properties or use network state.
      in: query
      name: mode
      example: INTRINSIC
      schema:
        $ref: "#/components/schemas/FeeEstimateMode"
    highVolumeThrottleQueryParam:
      description: The high-volume throttle utilization in basis points where 10000 equals 100%. A value of 0, the default, indicates no high-volume pricing.
      in: query
      name: high_volume_throttle
      example: 10000
      schema:
        type: integer
        minimum: 0
        maximum: 10000
        default: 0
    registeredNodeTypeQueryParam:
      example: BLOCK_NODE
      in: query
      name: type
      schema:
        $ref: "#/components/schemas/RegisteredNodeType"
...